hopefully the final update

+++ libft/ft_strncpy.c
@@ -6,13 +6,13 @@
 /*   By: rde-kwaa <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/19 11:15:31 by rde-kwaa          #+#    #+#             */
-/*   Updated: 2018/06/01 13:45:46 by rde-kwaa         ###   ########.fr       */
+/*   Updated: 2018/06/07 17:04:39 by rde-kwaa         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strncpy(char *des, char const *src, unsigned int n)
+char	*ft_strncpy(char *des, char const *src, size_t n)
 {
 	size_t	len;
 	size_t	i;
